moved files to submodule _gltf2usd
