+++ Source/_gltf2usd/gltf2/GLTFImage.py
