+++ Source/_gltf2usd/gltf2/Material.py
