+++ Source/_gltf2usd/gltf2/Mesh.py
