+++ Source/_gltf2usd/gltf2/Node.py
