+++ Source/_gltf2usd/gltf2/Scene.py
