+++ Source/_gltf2usd/gltf2/Skin.py
