+++ Source/_gltf2usd/gltf2loader.py
