+++ Source/_gltf2usd/gltf2usdUtils.py
