+++ Source/_gltf2usd/gltfUsdMaterialHelper.py
