+++ Source/_gltf2usd/usd_material.py
